backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.12.1
websockets>=12.0
google-genai==1.32.0
pydantic-settings>=2.0.0
